third_party/blink/web_tests/platform/generic/http/tests/security/sandboxed-iframe-blocks-access-from-parent-expected.txt

ALERT: PASS: Cross-origin access to the Location object threw an exception.
Sandboxing a frame puts it into a unique origin by default, which the containing document shouldn't have script access to. This test passes if an exception is thrown upon the access violation.

--------
Frame: 'testframe'
--------
